Initialize talk required python packages

Initialize talk required python packages to call from R.

Usage

talkrpp_initialize(
  python_executable = NULL,
  virtualenv = NULL,
  condaenv = "talkrpp_condaenv",
  ask = FALSE,
  refresh_settings = FALSE,
  save_profile = FALSE,
  check_env = TRUE,
  talkEmbed_test = FALSE,
  prompt = TRUE
)

Arguments

python_executable

the full path to the Python executable, for which talk required python packages is installed.

virtualenv

set a path to the Python virtual environment with talk required python packages installed Example: virtualenv = "~/myenv"

condaenv

set a path to the anaconda virtual environment with talk required python packages installed Example: condalenv = "myenv"

ask

logical; if FALSE, use the first talk required python packages installation found; if TRUE, list available talk required python packages installations and prompt the user for which to use. If another (e.g. python_executable) is set, then this value will always be treated as FALSE.

refresh_settings

logical; if TRUE, talk will ignore the saved settings in the profile and initiate a search of new settings.

save_profile

logical; if TRUE, the current talk required python packages setting will be saved for the future use.

check_env

logical; check whether conda/virtual environment generated by talkrpp_install() exists

talkEmbed_test

logical; Test whether function (talkEmbed) that requires python packages works.

prompt

logical; asking whether user wants to set the environment as default.